Fixed bugs in receiver tests.
