+++ SpeckleStructuralGSA.Test/SpeckleStructuralGSA.Test.csproj
@@ -160,6 +160,7 @@
     <Compile Include="Other\Result.cs" />
     <Compile Include="Other\SenderProcessor.cs" />
     <Compile Include="Tests\MeshTests.cs" />
+    <Compile Include="Tests\ModelValidationTests.cs" />
     <Compile Include="Tests\ReceiverTests.cs" />
     <Compile Include="Tests\SenderTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
